updates to the UI and updated logic to handle invalid access tokens in server

+++ main.js
@@ -44,20 +44,36 @@ function renderToken() {
 }
 
 function renderCart() {
-	let $container = $('.cart')
-	let cartItems = ""
-	cartItems = state.cart.map((item, index) => {
-			return `
+	const $container = $('.cart')
+	if (!state.cart[0]) {
+		const showCartButton = `<h6>Your cart is currently empty!</h6><button class="btn btn-primary btn-sm" id="get-cart" onclick="getCart()">Show Cart</button>`
+		$('.cart').html(showCartButton)
+		return
+	}
+	const hideCartButton = `<button class="btn btn-primary btn-sm" id="get-cart" onclick="hideCart()">Hide Cart</button>`
+	$container.html(hideCartButton)
+	const cartItems = state.cart.map((item, index) => {
+		return `
 				<li class="list-group-item" id="${index}"> ${item.name} - $${item.price}.00 <a class="text-danger" href="#" id="delete-${index}" onclick="deleteItem(${index})"><span class="oi oi-x text-danger ml-4" title="x" aria-hidden="true"></span></a></li>
 			`
-		})
-		cartItems.forEach((item) => $container.append(item))
-		const emptyCartButton = `<button class="btn btn-danger btn-sm mt-3" onclick="deleteCart()">Empty Cart</button>`
-		$container.append(emptyCartButton)
+	})
+	cartItems.forEach((item) => $container.append(item))
+	const cartPrices = state.cart.map((item) => {
+		return item.price
+	})
+	const cartTotal = cartPrices.reduce((accumulator, item) => accumulator + item)
+	$container.append(`<li class="list-group-item"><h5> Total: $${cartTotal}.00</h5></li>`)
+	const emptyCartButton = `<button class="btn btn-danger btn-sm" onclick="deleteCart()">Empty Cart</button>`
+	$container.append(emptyCartButton)
 }
 
 function getCart() {
 	let accessToken = state.token
+	if (!accessToken) {
+		const showCartButton = `<h6>You must login before you can view your cart</h6><button class="btn btn-primary btn-sm" id="get-cart" onclick="getCart()">Show Cart</button>`
+		$('.cart').html(showCartButton)
+		return
+	}
 	let url = `${BASEURL}/api/me/cart?accessToken=${accessToken}`
 	return fetch(url, {
 		method: 'GET',
@@ -80,14 +96,14 @@ function deleteItem(cartIndex) {
 	return fetch(url, {
 		method: 'DELETE',
 		headers: {
-			
+
 		}
 	})
-	.then(response => response.json())
-	.then((data) => {
-		getCart()
-	})
-	.catch(err => console.log(err))
+		.then(response => response.json())
+		.then((data) => {
+			getCart()
+		})
+		.catch(err => console.log(err))
 }
 
 function deleteCart() {
@@ -96,7 +112,7 @@ function deleteCart() {
 	return fetch(url, {
 		method: 'DELETE',
 		headers: {
-			
+
 		}
 	})
 		.then(response => response.json())
@@ -136,12 +152,53 @@ function getBrands() {
 		.catch(err => console.log(err))
 }
 
-function renderProducts () {
-	return
+function renderProducts() {
+	const $container = $('.products')
+	const productCards = state.products.map((item, index) => {
+		return `
+				<div class="col-md-6 mb-3 product-card" id="product-${index}"><div><img class="img-fluid" src="${item.imageUrls[0]}"></div><div class="text-center">${item.name}<button class="btn btn-outline-primary product-item mb-1" onclick="addProductToCart(${item.id})">Add to cart<br>$${item.price}.00</button></div></div>
+			`
+	})
+	productCards.forEach(product => $container.append(product))
 }
 
-function renderBrands () {
-	return
+function renderBrands() {
+	const $container = $('.brands')
+	const brandItems = state.brands.map((item, index) => {
+		return `
+				<li class="list-group-item list-group-item-action" id="brand-${index}"><a href="#"> ${item.name} </a></li>
+			`
+	})
+	brandItems.forEach((item) => $container.append(item))
+}
+
+function hideCart() {
+	let $container = $('.cart')
+	const showCartButton = `<button class="btn btn-primary btn-sm" id="get-cart" onclick="renderCart()">Show Cart</button>`
+	$container.html(showCartButton)
+}
+
+function addProductToCart(productId) {
+	let accessToken = state.token
+	if (!accessToken) {
+		const showCartButton = `<h6>You must login before you can add items to the cart</h6><button class="btn btn-primary btn-sm" id="get-cart" onclick="getCart()">Show Cart</button>`
+		$('.cart').html(showCartButton)
+		return
+	}
+	let url = `${BASEURL}/api/me/cart/${productId}?accessToken=${accessToken}`
+	console.log(url)
+	return fetch(url, {
+		method: 'POST',
+		headers: {
+			'content-type': 'application/json'
+		}
+	})
+		.then(response => response.json())
+		.then(data => {
+			state.cart = data
+			renderCart()
+		})
+		.catch(err => console.log(err))
 }
 
 getProducts()
